Extract Request and RequestStatusChange from model_soup
