+++ grouper/model_soup.py
@@ -22,17 +22,17 @@ from sqlalchemy.sql import label, literal
 from grouper.expiration import add_expiration, cancel_expiration
 from grouper.models.audit import Audit
 from grouper.models.audit_log import AuditLog
-from grouper.models.base.constants import OBJ_TYPES_IDX, REQUEST_STATUS_CHOICES
+from grouper.models.base.constants import OBJ_TYPES_IDX
 from grouper.models.base.model_base import Model
 from grouper.models.base.session import flush_transaction
 from grouper.models.comment import Comment, CommentObjectMixin
 from grouper.models.counter import Counter
 from grouper.models.permission import Permission
 from grouper.models.permission_map import PermissionMap
+from grouper.models.request import Request
+from grouper.models.request_status_change import RequestStatusChange
 from grouper.models.user import User
-from grouper.util import reference_id
 from .constants import MAX_NAME_LENGTH
-from .settings import settings
 
 GROUP_JOIN_CHOICES = {
     # Anyone can join with automatic approval
@@ -555,123 +555,6 @@ class Group(Model, CommentObjectMixin):
     def __repr__(self):
         return "<%s: id=%s groupname=%s>" % (
             type(self).__name__, self.id, self.groupname)
-
-
-class Request(Model, CommentObjectMixin):
-
-    __tablename__ = "requests"
-
-    id = Column(Integer, primary_key=True)
-
-    # The User that made the request.
-    requester_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    requester = relationship(
-        User, backref="requests", foreign_keys=[requester_id]
-    )
-
-    # The Group the requester is requesting access to.
-    requesting_id = Column(Integer, ForeignKey("groups.id"), nullable=False)
-    requesting = relationship(
-        Group, backref="requests", foreign_keys=[requesting_id]
-    )
-
-    # The User/Group which will become a member of the requested resource.
-    on_behalf_obj_type = Column(Integer, nullable=False)
-    on_behalf_obj_pk = Column(Integer, nullable=False)
-
-    edge_id = Column(Integer, ForeignKey("group_edges.id"), nullable=False)
-    edge = relationship("GroupEdge", backref="requests")
-
-    requested_at = Column(DateTime, default=datetime.utcnow, nullable=False)
-
-    status = Column(
-        Enum(*REQUEST_STATUS_CHOICES), default="pending", nullable=False
-    )
-
-    changes = Column(Text, nullable=False)
-
-    @property
-    def reference_id(self):
-        # type: () -> str
-        return reference_id(settings, "group", self)
-
-    def get_on_behalf(self):
-        obj_type = OBJ_TYPES_IDX[self.on_behalf_obj_type]
-
-        if obj_type == "User":
-            obj = User
-        elif obj_type == "Group":
-            obj = Group
-
-        return self.session.query(obj).filter_by(id=self.on_behalf_obj_pk).scalar()
-
-    def my_status_updates(self):
-
-        requests = self.session.query(
-            Request.id,
-            RequestStatusChange.change_at,
-            RequestStatusChange.from_status,
-            RequestStatusChange.to_status,
-            label("changed_by", User.username),
-            label("reason", Comment.comment)
-        ).filter(
-            RequestStatusChange.user_id == User.id,
-            Request.id == RequestStatusChange.request_id,
-            Comment.obj_type == 3,
-            Comment.obj_pk == RequestStatusChange.id,
-            Request.id == self.id
-        )
-
-        return requests
-
-    @flush_transaction
-    def update_status(self, requester, status, reason):
-        now = datetime.utcnow()
-        current_status = self.status
-        self.status = status
-
-        request_status_change = RequestStatusChange(
-            request=self,
-            user_id=requester.id,
-            from_status=current_status,
-            to_status=status,
-            change_at=now
-        ).add(self.session)
-        self.session.flush()
-
-        Comment(
-            obj_type=OBJ_TYPES_IDX.index("RequestStatusChange"),
-            obj_pk=request_status_change.id,
-            user_id=requester.id,
-            comment=reason,
-            created_on=now
-        ).add(self.session)
-
-        if status == "actioned":
-            edge = self.session.query(GroupEdge).filter_by(
-                id=self.edge_id
-            ).one()
-            edge.apply_changes(self)
-
-        Counter.incr(self.session, "updates")
-
-
-class RequestStatusChange(Model, CommentObjectMixin):
-
-    __tablename__ = "request_status_changes"
-
-    id = Column(Integer, primary_key=True)
-
-    request_id = Column(Integer, ForeignKey("requests.id"))
-    request = relationship(Request)
-
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    user = relationship(User, foreign_keys=[user_id])
-
-    from_status = Column(Enum(*REQUEST_STATUS_CHOICES))
-    to_status = Column(Enum(*REQUEST_STATUS_CHOICES), nullable=False)
-
-    change_at = Column(DateTime, default=datetime.utcnow, nullable=False)
 
 
 class GroupEdge(Model):
